Extract Jackson configuration to prevent dependency cycle
