+++ ngrinder-controller/src/main/java/org/ngrinder/infra/config/WebMvcConfig.java
@@ -1,6 +1,5 @@
 package org.ngrinder.infra.config;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
 import com.github.jknack.handlebars.springmvc.HandlebarsViewResolver;
 import org.ngrinder.infra.interceptor.DefaultSuccessJsonInterceptor;
 import org.ngrinder.infra.spring.ApiExceptionHandlerResolver;
@@ -8,7 +7,6 @@ import org.ngrinder.infra.spring.RemainedPathMethodArgumentResolver;
 import org.ngrinder.infra.spring.UserHandlerMethodArgumentResolver;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.autoconfigure.jackson.Jackson2ObjectMapperBuilderCustomizer;
 import org.springframework.boot.autoconfigure.web.ResourceProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
@@ -26,7 +24,6 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import org.springframework.web.servlet.i18n.LocaleChangeInterceptor;
 import org.springframework.web.servlet.mvc.method.annotation.ExceptionHandlerExceptionResolver;
 
-import java.text.SimpleDateFormat;
 import java.util.List;
 
 @Configuration
@@ -125,17 +122,5 @@ public class WebMvcConfig implements WebMvcConfigurer {
 			, "/**/*.html" , "/**/*.gif" , "/**/*.ico" , "/**/*.woff2"
 			, "/**/*.woff" , "/**/*.ttf"};
 		registry.addResourceHandler(staticPathPatterns).addResourceLocations(this.resourceProperties.getStaticLocations()).setCachePeriod(3600);
-	}
-
-
-	@Bean
-	public Jackson2ObjectMapperBuilderCustomizer jackson2ObjectMapperBuilderCustomizer() {
-		return customizer -> {
-			customizer.serializationInclusion(JsonInclude.Include.NON_NULL);
-			customizer.indentOutput(true);
-			customizer.timeZone("GMT");
-			customizer.dateFormat(new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'"));
-			customizer.modules(new NumberModule());
-		};
 	}
 }
