+++ ngrinder-controller/src/test/java/org/ngrinder/perftest/service/MockPerfTestRunnable.java
@@ -1,4 +1,4 @@
-/* 
+/*
  * Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
  *  You may obtain a copy of the License at
@@ -9,7 +9,7 @@
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
- * limitations under the License. 
+ * limitations under the License.
  */
 package org.ngrinder.perftest.service;
 
@@ -23,7 +23,6 @@ import org.ngrinder.infra.schedule.ScheduledTaskService;
 import org.ngrinder.infra.webhook.service.WebhookConfigService;
 import org.ngrinder.infra.webhook.service.WebhookService;
 import org.ngrinder.model.PerfTest;
-import org.springframework.context.annotation.Lazy;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
@@ -41,7 +40,7 @@ public class MockPerfTestRunnable extends PerfTestRunnable {
 								ConsoleManager consoleManager, PluginManager pluginManager,
 								Config config, ScheduledTaskService scheduledTaskService,
 								HazelcastService hazelcastService, AgentService agentService,
-								@Lazy WebhookService webhookService, WebhookConfigService webhookConfigService) {
+								WebhookService webhookService, WebhookConfigService webhookConfigService) {
 		super(perfTestService, agentManager, consoleManager, pluginManager, config,
 			scheduledTaskService, hazelcastService, agentService, webhookService, webhookConfigService);
 	}
